+++ infrastructure/terraform/per_account/dev/parameters/main.tf
@@ -62,3 +62,7 @@ resource "aws_secretsmanager_secret" "etl_notify_slack_webhook_url" {
 resource "aws_secretsmanager_secret" "apigee-app-client-info" {
   name = "${terraform.workspace}--apigee-app-client-info"
 }
+
+resource "aws_secretsmanager_secret" "apigee-sds-app-key" {
+  name = "${terraform.workspace}-apigee-sds-app-key"
+}