+++ infrastructure/terraform/per_account/int/parameters/main.tf
@@ -57,3 +57,7 @@ resource "aws_secretsmanager_secret" "ldap-changelog-password" {
 resource "aws_secretsmanager_secret" "etl_notify_slack_webhook_url" {
   name = "${terraform.workspace}--etl-notify-slack-webhook-url"
 }
+
+resource "aws_secretsmanager_secret" "apigee-sds-app-key" {
+  name = "${terraform.workspace}-apigee-sds-app-key"
+}