+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2024-09-02
+- [PI-302] SDS Smoke Test
+- [PI-245] Matrix Testing
+
 ## 2024-08-28
 - [PI-468] Bulk and upload fixes